setup-toolchain.sh: When translating -arch to -target, keep the original -arch as well
The clang driver does a number of rewrites back and forth between
arch names and triples. For aarch64/arm64, an input -arch arm64 or
-target arm64-apple-darwin ends up rewritten to aarch64-apple-darwin,
but at different stages (and -target aarch64-apple-darwin doesn't work
at all with current clang versions).

Keep the -arch argument, so that we follow the same path of arch/triple
rewrites as the original clang, so that -dumpmachine prints the same
value for -arch arm64 as it does on macos without this wrapper script
(printing aarch64-apple-darwin*).